+++ src/components/App.tsx
@@ -1,5 +1,5 @@
 import Page from "./Page.tsx";
-import { AuthProvider } from "./contexts/Providers.tsx";
+import AuthProvider from "./context/AuthProvider.tsx";
 
 export default function App() {
   return (
