+++ src/components/Cell.tsx
@@ -1,11 +1,11 @@
 import { useContext, useEffect, useState } from "react";
-import { parquetDataFrame } from "./tableProvider.js";
-import Layout from "./Layout.js";
+import { parquetDataFrame } from "../lib/tableProvider.js";
+import Layout from "./Layout.tsx";
 import { parquetMetadataAsync } from "hyparquet";
-import { NonHfUrl, FileUrl } from "./huggingface.ts";
+import { NonHfUrl, FileUrl } from "../lib/huggingface.ts";
 import Breadcrumb from "./Breadcrumb.tsx";
-import { asyncBufferFromUrl } from "./utils.ts";
-import { AuthContext } from "./contexts/authContext.ts";
+import { asyncBufferFromUrl } from "../lib/utils.ts";
+import { AuthContext } from "../contexts/authContext.ts";
 import { asyncRows } from 'hightable'
 
 interface CellProps {
