+++ src/components/Dropdown.tsx
@@ -1,5 +1,5 @@
 import { ReactNode, useEffect, useRef, useState } from 'react'
-import { cn } from './utils.js'
+import { cn } from '../lib/utils.js'
 
 interface DropdownProps {
   label?: string
