+++ src/components/Home.tsx
@@ -1,9 +1,9 @@
 import { useContext, useRef } from "react";
-import { changeQueryString } from "./huggingface.ts";
+import { changeQueryString } from "../lib/huggingface.ts";
 import Search from "./Search.tsx";
-import { login, logout } from "./login.ts";
-import { AuthContext } from "./contexts/authContext.ts";
-import HFLoginIcon from './assets/sign-in-with-huggingface-lg.svg';
+import { login, logout } from "../lib/login.ts";
+import { AuthContext } from "../contexts/authContext.ts";
+import HFLoginIcon from '../assets/sign-in-with-huggingface-lg.svg';
 
 /**
  * Home page
