+++ src/components/Link.tsx
@@ -1,4 +1,4 @@
-import { changeQueryString } from "./huggingface.ts";
+import { changeQueryString } from "../lib/huggingface.ts";
 import { ReactNode } from 'react'
 
 // workaround for Hugging Face spaces
