+++ src/components/LinkButton.tsx
@@ -1,4 +1,4 @@
-import { changeQueryString } from "./huggingface.ts";
+import { changeQueryString } from "../lib/huggingface.ts";
 import { ReactNode } from 'react'
 
 export default function LinkButton({ url, children }: { url: string; children?: ReactNode }) {
