+++ src/components/Main.tsx
@@ -1,7 +1,7 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 import App from './App.tsx'
-import './index.css'
+import '../styles/index.css'
 
 const root = document.getElementById('app')
 if (!root) throw new Error('No root element')
