+++ src/components/Search.tsx
@@ -1,8 +1,8 @@
 import { useContext, useEffect, useState } from "react";
 import { listDatasets, DatasetEntry } from "@huggingface/hub";
-import { baseUrl } from "./huggingface.ts";
+import { baseUrl } from "../lib/huggingface.ts";
 import Link from "./Link.tsx";
-import { AuthContext } from "./contexts/authContext.ts";
+import { AuthContext } from "../contexts/authContext.ts";
 
 export default function Search() {
   const [query, setQuery] = useState<string>();
