+++ src/components/viewers/ImageView.tsx
@@ -1,8 +1,8 @@
 import { useContext, useEffect, useState } from 'react'
 import { Spinner } from '../Layout.tsx'
 import ContentHeader from './ContentHeader.tsx'
-import { contentTypes, parseFileSize } from '../files.ts'
-import { AuthContext } from '../contexts/authContext.ts'
+import { contentTypes, parseFileSize } from "../../lib/files.ts"
+import { AuthContext } from "../../contexts/authContext.ts"
 
 enum LoadingState {
   NotLoaded,
