+++ src/lib/streamConverters.ts
