+++ src/styles/index.css
@@ -34,7 +34,7 @@
 }
 .brand::before {
   content: '';
-  background: url('assets/logo.svg') no-repeat 0 center;
+  background: url('../assets/logo.svg') no-repeat 0 center;
   background-size: 26px;
   height: 26px;
   width: 26px;
@@ -119,7 +119,7 @@ main {
   display: none;
 }
 .home {
-  background: url('assets/home.svg');
+  background: url('../assets/home.svg');
   background-position: left center;
   background-repeat: no-repeat;
   width: 22px;
@@ -146,7 +146,7 @@ main {
   }
 }
 .top-header .external {
-  background: url('assets/external-link.svg');
+  background: url('../assets/external-link.svg');
   background-position: center center;
   background-repeat: no-repeat;
   width: 32px;
@@ -176,7 +176,7 @@ main {
   overflow: hidden;
 }
 .branch-selector > button.dropdown-button::before {
-  background: url('assets/branch.svg') no-repeat center;
+  background: url('../assets/branch.svg') no-repeat center;
   background-size: 18px;
   position: static;
   content: "";
@@ -336,14 +336,14 @@ main {
 
 /* file icons */
 .file {
-  background: url(assets/file.svg);
+  background: url("../assets/file.svg");
   background-position: left center;
   background-repeat: no-repeat;
   background-size: 12px;
   padding-left: 22px;
 }
 .folder {
-  background-image: url(assets/folder.svg);
+  background-image: url("../assets/folder.svg");
 }
 
 /* viewer */
@@ -543,16 +543,16 @@ main {
   background-color: #cec;
 }
 .huggingface {
-  background: url('assets/huggingface.svg') no-repeat 8px center;
+  background: url('../assets/huggingface.svg') no-repeat 8px center;
 }
 .github {
-  background: url('assets/git.svg') no-repeat 8px center;
+  background: url('../assets/git.svg') no-repeat 8px center;
 }
 .aws {
-  background: url('assets/s3.svg') no-repeat 8px center;
+  background: url('../assets/s3.svg') no-repeat 8px center;
 }
 .azure {
-  background: url('assets/azure.svg') no-repeat 8px center;
+  background: url('../assets/azure.svg') no-repeat 8px center;
 }
 
 h1 {
