+++ tsconfig.app.tsbuildinfo
@@ -1,1 +1,1 @@
-{"root":["./src/App.tsx","./src/Breadcrumb.tsx","./src/Cell.tsx","./src/Dropdown.tsx","./src/File.tsx","./src/Folder.tsx","./src/Home.tsx","./src/Layout.tsx","./src/Link.tsx","./src/LinkButton.tsx","./src/Login.tsx","./src/Markdown.tsx","./src/Page.tsx","./src/Repository.tsx","./src/Search.tsx","./src/Viewer.tsx","./src/files.ts","./src/huggingface.spec.ts","./src/huggingface.ts","./src/main.tsx","./src/streamConverters.ts","./src/tableProvider.ts","./src/utils.ts","./src/vite-env.d.ts","./src/contexts/AuthContext.tsx","./src/viewers/ContentHeader.tsx","./src/viewers/ImageView.tsx","./src/viewers/MarkdownView.tsx","./src/viewers/ParquetView.tsx","./src/viewers/TextView.tsx","./src/workers/parquetWorker.ts","./src/workers/parquetWorkerClient.ts","./src/workers/sortParquetWorker.ts","./src/workers/sortParquetWorkerClient.ts","./src/workers/types.ts"],"version":"5.6.3"}
+{"root":["./src/vite-env.d.ts","./src/components/App.tsx","./src/components/Breadcrumb.tsx","./src/components/Cell.tsx","./src/components/Dropdown.tsx","./src/components/File.tsx","./src/components/Folder.tsx","./src/components/Home.tsx","./src/components/Layout.tsx","./src/components/Link.tsx","./src/components/LinkButton.tsx","./src/components/Main.tsx","./src/components/Markdown.tsx","./src/components/Page.tsx","./src/components/Search.tsx","./src/components/Viewer.tsx","./src/components/context/AuthProvider.tsx","./src/components/viewers/ContentHeader.tsx","./src/components/viewers/ImageView.tsx","./src/components/viewers/MarkdownView.tsx","./src/components/viewers/ParquetView.tsx","./src/components/viewers/TextView.tsx","./src/contexts/authContext.ts","./src/lib/files.ts","./src/lib/huggingface.ts","./src/lib/login.ts","./src/lib/streamConverters.ts","./src/lib/tableProvider.ts","./src/lib/utils.ts","./src/test/lib/huggingface.spec.ts","./src/workers/parquetWorker.ts","./src/workers/parquetWorkerClient.ts","./src/workers/sortParquetWorker.ts","./src/workers/sortParquetWorkerClient.ts","./src/workers/types.ts"],"version":"5.6.3"}
